ActiveSupport::Cache::MonetaStore: fix writing raw values
When the :raw option is used, the plain string value of the entry is
written to the store.  This is important for using `#increment` and
`#decrement` with an existing value, which is also the only way to use
these methods in a way that is consistent with other
ActiveSupport::Cache stores.

This commit also updates the specs for this to help ensure that the
MonetaStore behavior matches that of the MemoryStore and
MemcacheStore (bundled with ActiveSupport), by running the same
specs - with some exceptions - against each implementation.