Moneta::Shared: recover if server instance is closed
When using the Shared wrapper, the first created instance of the store
rusn a server, and all other instances run a client that connects to
that server.  If the first instance is closed, the server stops and the
other instances start failing.  This fix allows the other instances to
attempt to launch the server themselves once they notice that it is
down.

A mutex is added so that in multithreaded environments, accessing the
store from different threads does not trigger any races while the server
is launching.