Add `not_supports` class method to stores
This method can be used to subtract features from a store, which can be
useful in some cases (e.g. with the `each_key` method).  In proxies,
using `not_supports` can be used to mask a proxied method.

Key transformers now use this method to block access to the underlying
store's each_key method.  This is done because we don't have an easy way
of reversing the key transform at present