specs: add micosecond-aware testing for expiry
Some adapters have native expiry features that fail if the test is
performed when the clock is at a particular microsecond
position within a second (usually towards the end of the second).  The
main offender in this regard is memcached; see
https://github.com/memcached/memcached/issues/307.

This patch makes testing of expiry functionality configurable for each
adapter, and adds tests to make sure that expiry works at different
microsecond positions.