allow overriding socket in test suite
Debian has been carrying a patch to allow for tests to run on the
MySQL socket instead of localhost for a while, but has never been
proposed upstream. This is an attempt at forward-porting this patch
from 1.0 to 1.5.

The reason we do this is (presumably!) that our test suites might run
without network and it seems more reliable to test on the socket.